+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 * Add new command :command:`call-phenotypes`.
 * Add ``--phenotypes`` argument  to :command:`combine-results` command.
 * Deprecate :meth:`api.utils.load_control_table` method.
+* Split ``api.utils`` submodule into two submodules ``api.utils`` and ``api.core``.
 
 0.5.0 (2021-10-02)
 ------------------